requirements.txt
# Core
flask>=3.0.0
flask-sqlalchemy>=3.1.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
# SQLite is built-in, used as fallback

# Exchange APIs
ccxt>=4.2.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# HTTP Client
requests>=2.31.0
aiohttp>=3.9.0

# AI Engine
openai>=1.0.0

# Scheduling
apscheduler>=3.10.0

# Async
asyncio-throttle>=1.0.2
